% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizards.R
\name{wizard_import}
\alias{wizard_import}
\title{Guided Import Wizard}
\usage{
wizard_import()
}
\value{
A \code{moover_spec} object.
}
\description{
Interactively builds a simple import spec, previews the canonicalised data,
and writes the run spec into the current workspace.
}
