% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizards.R
\name{wizard_predict}
\alias{wizard_predict}
\title{Guided Prediction Wizard}
\usage{
wizard_predict()
}
\value{
A \code{moover_spec} object.
}
\description{
Builds a simple prediction spec for applying an existing model bundle to new
data.
}
